Move to own test class
